+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Procedure when bumping the version number:
 
 ### Added
 - `Frame::set_title`
+- `WidgetExt::title`
+- `widgets::title`
 
 ## v0.2.0 - 2023-08-31